refactor(ui): extract settings/data/form-state.ts — formStateFromPayload + ConfigPayload
